Implement updated API
