+++ .gitmodules
@@ -1,8 +1,4 @@
-[submodule "src/main/resources/migration"]
-	path = src/main/resources/migration
-	url = https://github.com/OpenAlcoholics/CockBot-OpenAPI.git
-	branch = master
 [submodule "src/main/resources/openapi"]
 	path = src/main/resources/openapi
 	url = https://github.com/OpenAlcoholics/CockBot-OpenAPI.git
-	branch = feature/opencocktailapi
+	branch = develop
