+++ src/main/java/group/openalcoholics/cocktailparty/api/Api.kt
@@ -7,6 +7,7 @@ import group.openalcoholics.cocktailparty.api.handler.CocktailHandler
 import group.openalcoholics.cocktailparty.api.handler.GlassHandler
 import group.openalcoholics.cocktailparty.api.handler.IngredientCategoryHandler
 import group.openalcoholics.cocktailparty.api.handler.IngredientHandler
+import group.openalcoholics.cocktailparty.api.handler.SecurityHandler
 import group.openalcoholics.cocktailparty.api.handler.VersionHandler
 import group.openalcoholics.cocktailparty.module.ApiConfig
 import group.openalcoholics.cocktailparty.module.AuthConfig
@@ -14,12 +15,7 @@ import io.vertx.core.AbstractVerticle
 import io.vertx.core.Future
 import io.vertx.core.http.HttpServerOptions
 import io.vertx.core.json.Json
-import io.vertx.ext.auth.jwt.JWTAuth
 import io.vertx.ext.web.api.contract.openapi3.OpenAPI3RouterFactory
-import io.vertx.kotlin.core.json.json
-import io.vertx.kotlin.core.json.obj
-import io.vertx.kotlin.ext.auth.PubSecKeyOptions
-import io.vertx.kotlin.ext.auth.jwt.JWTAuthOptions
 import io.vertx.kotlin.ext.web.api.contract.RouterFactoryOptions
 import mu.KotlinLogging
 
@@ -62,34 +58,7 @@ class Api @Inject constructor(
                     .register(cocktailCategoryHandler)
                     .register(cocktailHandler)
 
-                val sha = authConfig.sha
-                if (sha !in listOf(256, 384, 512))
-                    throw ApiInitializationException("auth.sha must be one of [256, 384, 512]")
-
-                val jwtOptions = JWTAuthOptions(
-                    pubSecKeys = listOf(PubSecKeyOptions(
-                        algorithm = "RS$sha",
-                        publicKey = authConfig.publicKey)))
-                val jwtAuth = JWTAuth.create(vertx, jwtOptions)!!
-
-                routerFactory.addSecurityHandler("Token") { ctx ->
-                    val rawToken: String? = ctx.request().getHeader("Authorization")
-                    if (rawToken == null
-                        || !rawToken.startsWith("Bearer ")) ctx.fail(Status.UNAUTHORIZED)
-                    else {
-                        val token = rawToken.substring("Bearer ".length)
-                        jwtAuth.authenticate(json {
-                            obj("jwt" to token)
-                        }) { result ->
-                            if (result.succeeded()) {
-                                ctx.setUser(result.result())
-                                ctx.next()
-                            } else {
-                                ctx.fail(Status.UNAUTHORIZED)
-                            }
-                        }
-                    }
-                }
+                routerFactory.addSecurityHandler("Token", SecurityHandler(vertx, authConfig))
 
                 val router = routerFactory.router!!
